Move common PZEM logic to base class
