+++ include/pzems/acpzem.h
@@ -5,16 +5,9 @@
 #include <Arduino.h>
 #include <ArduinoJson.h>
 #include <PZEM004Tv30.h>
-#include "utils/date.h"
+#include "pzem.h"
 
-struct Zone {
-  float t1StartEnergy;
-  float t2StartEnergy;
-  float t1EnergyAcc;
-  float t2EnergyAcc;
-};
-
-class AcPzem {
+class AcPzem : public Pzem {
  public:
   AcPzem(SoftwareSerial& port, uint8_t addr = PZEM_DEFAULT_ADDR);
 
@@ -26,32 +19,14 @@ class AcPzem {
 
  private:
   PZEM004Tv30 _pzem;
-  Zone _zone;
 
   float _voltage;
   float _current;
   float _power;
-  float _energy;
   float _frequency;
   float _powerFactor;
-  float _t1Energy;
-  float _t2Energy;
-  Date _createdAt;
 
   void readValues();
-
-  void calcZoneEnergy();
-
-  float calcT1ZoneEnergy();
-  float calcT2ZoneEnergy();
-
-  bool isT1ZoneActive(uint8_t hour);
-
-  bool isStartOfT1Zone(uint8_t hour, uint8_t minute, uint8_t second);
-  bool isEndOfT1Zone(uint8_t hour, uint8_t minute, uint8_t second);
-
-  bool isStartOfT2Zone(uint8_t hour, uint8_t minute, uint8_t second);
-  bool isEndOfT2Zone(uint8_t hour, uint8_t minute, uint8_t second);
 };
 
 #endif
